move SuiteContext out of `_test` namespace
This allows it to be easily re-used in other libraries to do
meta-testing on their own godog features/steps.

At some point it may make sense to migrate this entirely to a
sub-package of godog, however that would require a larger
re-architecting of the project structure to avoid circular dependencies
(e.g. `import cycle not allowed`).
